refactor: separate input comment
