+++ src/types/index.tsx
@@ -1,0 +1,8 @@
+interface CommentDataType {
+  photoUrl: string;
+  username: string;
+  content: string;
+  time: string;
+}
+
+export type { CommentDataType };
